ftr: file output without removing on error
The standard CL behavior is to remove files that contain partial
output (when using with-open-file).  This is often not what I want.  The
macros WITH-FILE-OUTPUT and WITH-FILES-FOR-OUTPUT will open files that
will be closed, but not removed (e.g. left in a partially written state)
if the computation is aborted via a restart.